Under the folder Database Connector, there were three files db_bingcovid.py, db_maylasia_states.py and db_malaysia_patient_cases.py. Three of these files were using similar functions of connecting the database on mySQL. Using Factory Method I created one file with three different instantces of Factory_Demo with initial common functions and then their own implementation. Creational Design pattern here was used to showcase polymorphism and to avoid complexity of three different files where my one main class created instances to me implemented by 3 different classes.
